Splitter aarsavregningUtenEllerDeltGrunnlag.tsx til to filer, aarsavregningUtenEllerDeltGrunnlag.tsx laster inn data, oppretter medlemskapsperioder dersom delt grunnlag og gjør klar default verdier for ny komponent aarsavregningFormComponent.tsx.
